+++ apps/client/src/app/components/admin-overview/admin-overview.component.ts
@@ -50,7 +50,7 @@ export class AdminOverviewComponent implements OnDestroy, OnInit {
     private dataService: DataService,
     private userService: UserService
   ) {
-    this.info = this.dataService.fetchInfo();
+    this.info = this.dataService.getInfo();
 
     this.userService.stateChanged
       .pipe(takeUntil(this.unsubscribeSubject))