+++ apps/client/src/app/pages/features/features-page.component.ts
@@ -23,7 +23,7 @@ export class FeaturesPageComponent implements OnDestroy {
     private dataService: DataService,
     private userService: UserService
   ) {
-    this.info = this.dataService.fetchInfo();
+    this.info = this.dataService.getInfo();
   }
 
   public ngOnInit() {