+++ apps/client/src/app/pages/home/home-page.component.ts
@@ -35,7 +35,7 @@ export class HomePageComponent implements OnDestroy, OnInit {
     private dataService: DataService,
     private userService: UserService
   ) {
-    this.info = this.dataService.fetchInfo();
+    this.info = this.dataService.getInfo();
 
     this.userService.stateChanged
       .pipe(takeUntil(this.unsubscribeSubject))