+++ apps/client/src/app/pages/register/register-page.component.ts
@@ -39,13 +39,13 @@ export class RegisterPageComponent implements OnDestroy, OnInit {
     private tokenStorageService: TokenStorageService,
     private userService: UserService
   ) {
-    this.info = this.dataService.fetchInfo();
+    this.info = this.dataService.getInfo();
 
     this.tokenStorageService.signOut();
   }
 
   public ngOnInit() {
-    const { demoAuthToken, globalPermissions } = this.dataService.fetchInfo();
+    const { demoAuthToken, globalPermissions } = this.dataService.getInfo();
 
     this.demoAuthToken = demoAuthToken;
     this.deviceType = this.deviceService.getDeviceInfo().deviceType;