+++ libs/common/src/lib/exceptions/not-found-error.ts
@@ -1,0 +1,6 @@
+export class NotFoundError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = 'NotFoundError';
+  }
+}