+++ CHANGELOG.md
@@ -36,6 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added an error message for duplicates to the preview step of the activities import
 - Added a connection timeout to the environment variable `DATABASE_URL`
 - Introduced the _Open Startup_ (`/open`) page with aggregated key metrics including uptime
+- POST api to add benchmarks
+- Ability to add symbol as benchmark from Market Data
 
 ### Changed